Shopping flow is more or less done

+++ server/src/main/java/agh/queueFreeShop/service/ShopService.java
@@ -1,17 +1,20 @@
 package agh.queueFreeShop.service;
 
+import agh.queueFreeShop.exception.ForbiddenException;
 import agh.queueFreeShop.exception.NotFoundException;
-import agh.queueFreeShop.model.CartItem;
-import agh.queueFreeShop.model.Product;
-import agh.queueFreeShop.model.Receipt;
-import agh.queueFreeShop.model.ShoppingCart;
+import agh.queueFreeShop.model.*;
+import agh.queueFreeShop.physical.EntranceWeight;
+import agh.queueFreeShop.physical.ExitWeight;
 import agh.queueFreeShop.repository.CartItemRepository;
-import agh.queueFreeShop.repository.ProductRepository;
 import agh.queueFreeShop.repository.ReceiptRepository;
 import agh.queueFreeShop.repository.ShoppingCartRepository;
+import agh.queueFreeShop.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.HashSet;
+
+import static java.lang.Math.abs;
 
 /**
  * Responsible for business logic - all transactions are executed there.
@@ -22,11 +25,21 @@ public class ShopService {
     private final ShoppingCartRepository cartRepository;
     private final ReceiptRepository receiptRepository;
     private final CartItemRepository cartItemRepository;
+    private final UserRepository userRepository;
 
-    ShopService(ShoppingCartRepository cartRepository, ReceiptRepository receiptRepository, CartItemRepository cartItemRepository) {
+    private final EntranceWeight entranceWeight;
+    private final ExitWeight exitWeight;
+
+    private User enteringCustomer;
+    private User leavingCustomer;
+
+    ShopService(ShoppingCartRepository cartRepository, ReceiptRepository receiptRepository, CartItemRepository cartItemRepository, UserRepository userRepository, EntranceWeight entranceWeight, ExitWeight exitWeight) {
         this.cartRepository = cartRepository;
         this.receiptRepository = receiptRepository;
         this.cartItemRepository = cartItemRepository;
+        this.userRepository = userRepository;
+        this.entranceWeight = entranceWeight;
+        this.exitWeight = exitWeight;
     }
 
     /**
@@ -55,8 +68,7 @@ public class ShopService {
             if (cartItem.getQuantity() == 1) {
                 cart.getItems().remove(cartItem);
                 cartItemRepository.delete(cartItem);
-            }
-            else
+            } else
                 cartItem.removeOne();
         }
 
@@ -66,7 +78,69 @@ public class ShopService {
     @Transactional
     public Receipt finalizeShopping(ShoppingCart cart) {
         Receipt receipt = cart.generateReceipt();
-        cartRepository.delete(cart);
+        cart.setFinalized(true);
+        cartRepository.save(cart);
         return receiptRepository.save(receipt);
+    }
+
+    public void onScannedEnteringCustomer(Long userId) {
+        User user = userRepository.getById(userId);
+        enteringCustomer = user;
+
+        if (user == null)
+            throw new NotFoundException("User not found");
+
+        sendNotification(user);
+    }
+
+    public void onScannedLeavingCustomer(Long userId) {
+        User user = userRepository.getById(userId);
+        leavingCustomer = user;
+
+        if (user == null)
+            throw new NotFoundException("User not found");
+
+        sendNotification(user);
+    }
+
+    @Transactional
+    public ShoppingCart onCustomerConfirmedEntry(User user) {
+        if (enteringCustomer == null || user.getId() != enteringCustomer.getId())
+            throw new ForbiddenException("Customer not at entrance");
+
+        ShoppingCart cart = new ShoppingCart();
+        cart.setUser(user);
+
+        int weight = entranceWeight.readWeight();
+        cart.setInitialWeight(weight);
+
+        cart.setItems(new HashSet<>());
+        cart.setFinalized(false);
+        cart = cartRepository.save(cart);
+        return cart;
+    }
+
+    @Transactional
+    public void onCustomerConfirmedExit(User user) {
+        if (leavingCustomer == null || user.getId() != leavingCustomer.getId())
+            throw new ForbiddenException("Customer not at exit");
+
+        ShoppingCart cart = cartRepository.getByUserId(user.getId());
+
+        int finalWeight = exitWeight.readWeight();
+
+        if (!validateWeight(cart, finalWeight))
+            throw new ForbiddenException("Final weight is incorrect");
+
+        cartRepository.delete(cart);
+    }
+
+    public boolean validateWeight(ShoppingCart cart, int finalWeight) {
+        int expectedWeight = cart.getProductsWeight() + cart.getInitialWeight();
+        return abs(finalWeight - expectedWeight) < 1000;
+    }
+
+    public void sendNotification(User user) {
+        //TODO
     }
 }
